+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ Each individual change should have a link to the pull request after the descript
 
 Added
 ^^^^^
+- added support for weights to mean, median and mode imputers `#47 <https://github.com/lvgig/tubular/pull/47>`_
+
 - added classname() method to BaseTransformer and prefixed all errors with classname call for easier debugging `#48 <https://github.com/lvgig/tubular/pull/48>`_
 
 Fixed